Add mx_fail_domains option to the dnslookup router
There seems to be a bug in exim whereby it fails to resolve domains for
which there exists an A record but no MX record. In such cases DNS
servers will respond with a NODATA response, which has a status of
NOERROR and an SOA record in the AUTHORITY section. Exim seems to think
that if there's no error code the response should either be an MX record
or a CNAME, which it isn't so returns DNS_FAIL.

The `mx_fail_domains` option tells Exim to treat MX failures as NODATA
responses and proceed with AAAA and A record lookups.

Add this option to the dnslookup router and make it settable with an
environment variable. If no value is set it will have no effect.