Change base image to ubuntu:rolling
Alpine uses musl-libc as its standard C library instead of glibc like
other distros. The res_search() in musl-libc is subtly different to the
one in glibc in that it does not return a return code of -1 in the event
of a DNS lookup error but instead returns the length of the response.
This breaks Exim's dns_basic_lookup function, resulting in Exim failing
to resolve domains that have an A record but no MX.

Change the base image to ubuntu:rolling to get a libc Exim works with.
The paths all change to refer to exim4.